# api-spec/openapi.yaml
openapi: 3.0.0
info:
  title: SMS Gateway
  description: ""
  contact:
    email: petretiandrea@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://agenda.beaesthetic.it/
  - url: https://localhost:8080/
paths:
...
  /messages/{smsId}:
    get:
      tags:
        - sms
      summary: "Get an sms"
      operationId: getSmsById
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: smsId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Sms"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsEntityResponse"
        "404":
          description: "Sms not found"
  /messages:
    get:
      tags:
        - sms
      summary: "Get all filtered sms"
      operationId: getMessages
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
        - in: query
          name: isSent
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: "Found messges"
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/SmsEntityResponse"
        '400':
          description: "Invalid request"
    post:
      tags:
        - sms
      summary: "Send a new sms"
      operationId: sendSms
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: header
          name: "Idempotency-Key"
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSmsRequest"
      callbacks:
        messageCallback:
          '/callbackUrl':
            post:
              operationId: notifyMessageDelivery
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/EventNotificationDto'
              responses:
                200:
                  description: "Callback accepted"
      responses:
        "201":
          description: "Send sms takes"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsEntityResponse"
        '400':
          description: "Invalid request"
...
components:
  schemas:
...
    FailedAttempt:
      type: object
      properties:
        type:
          type: string
          format: "failure"
        reason:
          type: string
          description: "Reason of failure attempt"
      required:
        - type
    SuccessfulAttempt:
      type: object
      properties:
        type:
          type: string
          pattern: "success"
      required:
        - type
...
    EventNotificationType:
      type: string
      enum:
        - message.deliver.succeeded
        - message.deliver.failed
    EventNotificationDto:
      type: object
      properties:
        eventType:
          $ref: '#/components/schemas/EventNotificationType'
        data:
          $ref: '#/components/schemas/SmsEntityResponse'
        metadata:
          $ref: '#/components/schemas/AdditionalMetadata'
      required:
        - eventType
        - data
        - metadata
    SmsEntityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        from:
          type: string
        to:
          type: string
        owner:
          type: string
          format: uuid
        isSent:
          type: boolean
        lastAttempt:
          type: object
          properties:
            attemptCount:
              type: integer
          oneOf:
            - $ref: '#/components/schemas/SuccessfulAttempt'
            - $ref: '#/components/schemas/FailedAttempt'
          discriminator:
            propertyName: type
            mapping:
              success: '#/components/schemas/SuccessfulAttempt'
              failure: '#/components/schemas/FailedAttempt'
          required:
            - attemptCount
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - from
        - to
        - owner
        - isSent
        - createdAt
        - updatedAt
    SendSmsRequest:
      type: object
      properties:
        content:
          type: string
        from:
          type: string
        to:
          type: string
        webhook:
          $ref: '#/components/schemas/WebhookOnDemandConfig'
        metadata:
          $ref: '#/components/schemas/AdditionalMetadata'
      required:
        - content
        - from
        - to
    WebhookOnDemandConfig:
      type: object
      properties:
        url:
          type: string
    AdditionalMetadata:
      type: object
      additionalProperties:
        type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Api-Key